Fixing get & post
