+++ server.js
@@ -1,0 +1,19 @@
+const express = require('express');
+
+const HubsRouter = require('./hubs/hubs-router');
+
+const server = express();
+
+server.use(express.json());
+server.use('/api/hubs', HubsRouter);
+
+server.get('/', (req, res) => {
+  res.send(`
+    <h2>Lambda Hubs API</h>
+    <p>Welcome to the Lambda Hubs API</p>
+  `);
+});
+
+
+
+module.exports = server;
